@@ -19,7 +19,7 @@ package_dir =
 packages =
     grizzly_reduce_monitor
 zip_safe = False
-python_requires = >=3.8
+python_requires = >=3.10
 
 [options.entry_points]
 console_scripts =
