@@ -1,9 +1,7 @@
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file, You can
 # obtain one at http://mozilla.org/MPL/2.0/.
-"""Launcher to redirect all stdout/stderr to a private log file.
-"""
-
+"""Launcher to redirect all stdout/stderr to a private log file."""
 
 import argparse
 import ctypes
@@ -11,7 +9,6 @@ import os
 import sys
 from logging import getLogger
 from pathlib import Path
-from typing import List, Optional
 
 from .common import CommonArgParser, ReductionWorkflow
 
@@ -21,13 +18,13 @@ LOG = getLogger(__name__)
 class PrivateLogLauncher(ReductionWorkflow):
     """Launcher for a fuzzing pool, using docker parameters from a private repo."""
 
-    def __init__(self, command: List[str], log_dir: Path) -> None:
+    def __init__(self, command: list[str], log_dir: Path) -> None:
         super().__init__()
         self.command = command.copy()
         self.environment = os.environ.copy()
         self.log_dir = log_dir
 
-    def run(self) -> Optional[int]:
+    def run(self) -> int | None:
         assert self.command
 
         LOG.info("Creating private logs directory '%s/'", self.log_dir)
@@ -50,7 +47,7 @@ class PrivateLogLauncher(ReductionWorkflow):
         os.execvpe(self.command[0], self.command, self.environment)
 
     @staticmethod
-    def parse_args(args: Optional[List[str]] = None) -> argparse.Namespace:
+    def parse_args(args: list[str] | None = None) -> argparse.Namespace:
         parser = CommonArgParser(prog="grizzly-reduce-tc-log-private")
         parser.add_argument(
             "--log-dir",
